+++ CHANGELOG.md
@@ -5,8 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-
 ## [UNRELEASED]
+
+### Added
+
+- Implement `Field` question type for new GLPI forms
 
 ## [1.22.2] - 2025-10-24